test: split <checkboxGroup> component tests to checkboxgroup.spec.js
