+++ packages/checkbox/src/__tests__/checkbox.spec.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { mount, shallow } from 'enzyme'
-import { Checkbox, CheckboxGroup } from '../react-component-node-modules/index'
+import { Checkbox } from '../react-component-node-modules/index'
 import { configure } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
 
@@ -22,35 +22,14 @@ it('render <Checkbox> component', () => {
   expect(checkboxCheckedDisabled).toMatchSnapshot()
 })
 
-it('render <CheckboxGroup> component', () => {
-  const checkboxGroup = shallow(
-    <CheckboxGroup
-      options={['Apple', 'Banana', 'Cat']}
-      span={30}
-      theme="light"
-      direction="row"
-    />,
-  )
-  expect(checkboxGroup).toMatchSnapshot()
-})
-
 it('renders without crashing', () => {
   const div = document.createElement('div')
   ReactDOM.render(<Checkbox>Test</Checkbox>, div)
-  ReactDOM.render(
-    <CheckboxGroup
-      options={['Apple', 'Banana', 'Cat']}
-      span={30}
-      theme="light"
-      direction="row"
-    />,
-    div,
-  )
 })
 
 it('checkbox image changed after clicking label', () => {
   const checkbox = mount(<Checkbox>Test</Checkbox>)
-  
+
   expect(checkbox.find('img').prop('src')).toEqual('btn_checkbox.svg')
   checkbox.find('input').simulate('click')
   expect(checkbox.find('img').prop('src')).toEqual('btn_checkbox_pressed.svg')
@@ -82,17 +61,4 @@ it('renders an image with src correctly', () => {
   expect(checkboxCheckedDisabled.find('img').prop('src')).toEqual(
     'btn_checkbox_presseddisable.svg',
   )
-})
-
-it('render CheckboxGroup correctly', () => {
-  const checkboxGroup = mount(
-    <CheckboxGroup
-      options={['Apple', 'Banana', 'Cat']}
-      span={30}
-      theme="light"
-      direction="row"
-    />,
-  )
-  expect(checkboxGroup.find('span').length).toEqual(3)
-  expect(checkboxGroup.children().text()).toEqual('AppleBananaCat')
 })
